Delete mocks from system module and move them in test-utils
